requirements.txt
customtkinter==5.1.2
Pillow==9.5.0
Pmw==2.1.1
pytube==12.1.3
Requests==2.29.0
spotipy==2.23.0
